Moved dataprocessing code to its own package
